Using components.
