+++ Introduction-to-components/components/src/App.vue
@@ -1,5 +1,5 @@
 <template>
-  <app-server-status></app-server-status>
+  <app-server></app-server>
 </template>
 
 <script>
