+++ Introduction-to-components/components/src/Home.vue
@@ -1,19 +1,13 @@
 <template>
   <div>
-    <p>Server Status: {{status}}</p>
-    <hr />
-    <button @click="changeStatus">Change Status</button>
+    <app-server-status v-for="server in 5"></app-server-status>
   </div>
 </template>
 <script>
+import ServerStatus from "./ServerStatus.vue";
 export default {
-  data: function() {
-    status: "Critical";
-  },
-  methods: {
-    changeStatus() {
-      this.status = "Normal";
-    }
+  components: {
+    "app-server-status": ServerStatus
   }
 };
 </script>
