+++ Introduction-to-components/components/src/ServerStatus.vue
@@ -1,0 +1,19 @@
+<template>
+  <div>
+    <p>Server Status: {{status}}</p>
+    <hr />
+    <button @click="changeStatus">Change Status</button>
+  </div>
+</template>
+<script>
+export default {
+  data: function() {
+    status: "Critical";
+  },
+  methods: {
+    changeStatus() {
+      this.status = "Normal";
+    }
+  }
+};
+</script>
