+++ Introduction-to-components/components/src/main.js
@@ -2,7 +2,7 @@ import Vue from "vue";
 import App from "./App.vue";
 import Home from "./Home.vue";
 
-Vue.component("app-server-status", Home);
+Vue.component("app-server", Home);
 
 new Vue({
   el: "#app",
